feat(import): setup tabs for settings
